Add aggresively optimized single-float versions
Sometimes you just want the speed. This branch throws safety to the wind
and gives you that.

Originally this was going to be integrated with the defeasing macro
however there were enough little changes that meant it was better to
seperate the implementations rather than make the original version ugly.