Change to non-import api
This is a personal preference thing but for simple maths functions I
prefer to use an API that you don't `:use` in your package. Instead you
reference directly using the package name.

This means you don't add a pile of symbols to your package and you can
use shorter names (as they won't collide)

For example:

```
   CL-USER> (easing:in-cubic 0.3)
   0.027000003
   CL-USER> (easing-f:in-cubic 0.3)
   0.027000003
```